Changed, use off64_t for seeks. The size of off_t is platform dependent and always needs diddling to get >2GB file support working on 32-bit platforms

+++ src/smb_file.c
@@ -282,7 +282,7 @@ ssize_t   smb_fwrite(smb_session *s, smb_fd fd, void *buf, size_t buf_size)
     return resp->data_len;
 }
 
-ssize_t   smb_fseek(smb_session *s, smb_fd fd, off_t offset, int whence)
+off64_t   smb_fseek(smb_session *s, smb_fd fd, off64_t offset, int whence)
 {
     smb_file  *file;
